+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/docs
 
+## 2.7.29
+
+### Patch Changes
+
+- Various changes
+
+## 2.7.28
+
 ## 2.7.27
 
 ### Patch Changes