+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-ui
 
+## 2.7.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.29
+
+## 2.7.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.28
+
 ## 2.7.27
 
 ### Patch Changes